Update project version from 6.4.2 to 6.4.3

+++ hosting/build.ps1
@@ -1,24 +1,29 @@
+param(
+	[switch]
+	[bool]$alias
+)
 $InformationPreference = "Continue";
 $install = $env:InstallDirectory;
 
+if(-not $alias) {
+	$projects = @(
+		"sample.hosting.utility"
+		,"sample.hosting.webapi"
+		,"sample.hosting.daemon"
+	);
 
-$projects = @(
-	"sample.hosting.utility"
-	,"sample.hosting.webapi"
-	,"sample.hosting.daemon"
-);
-
-foreach($item in $projects){
-	if(Test-Path "$install/$item" -type Container){
-		Write-Information "Deleting $item";
- 		Get-ChildItem $install\$item | Remove-Item -Recurse -Force;
+	foreach($item in $projects){
+		if(Test-Path "$install/$item" -type Container){
+			Write-Information "Deleting $item";
+ 			Get-ChildItem $install\$item | Remove-Item -Recurse -Force;
+		}
 	}
-}
 
-dotnet restore $PSScriptRoot --interactive;
-foreach($project in $projects){
-	"Building $project";
-	dotnet publish $PSScriptRoot\$project\$project.csproj -o $install\$project -c debug --interactive;
+	dotnet restore $PSScriptRoot --interactive;
+	foreach($project in $projects){
+		"Building $project";
+		dotnet publish $PSScriptRoot\$project\$project.csproj -o $install\$project -c debug --interactive;
+	}
 }
 
 set-alias -name sample-hosting -Value $env:InstallDirectory\sample.hosting.utility\sample.hosting.utility.exe
